pyproject.toml
[tool.poetry]
name = "wpp_bot"
version = "1"
description = "Sends messages with whatsapp web from a list of numbers obtained with google sheets data"
authors = ["rajmaninov1 <8.8tolosajulian@gmail.com>"]
license = "MIT License"

[tool.poetry.dependencies]
python = "^3.10"
google-auth = "^2.6.6"
google-auth-oauthlib = "^0.5.1"
gspread = "^5.4.0"
oauth2client = "^4.1.3"
oauthlib = "^3.2.0"
selenium = "^4.3.0"
webdriver-manager = "^3.8.3"
phonenumbers = "^8.12.53"
alright = {git = "https://github.com/Kalebu/alright.git", rev = "main"}
fastapi = "^0.85.1"
fake-http-header = "^0.1.2"
gunicorn = "^20.1.0"
uvicorn = "^0.19.0"
packaging = "^21.3"
bs4 = "^0.0.1"
pydantic = "^1.10.2"
sqlalchemy = "^1.4.44"
fastapi-sqlalchemy = "^0.2.1"
alembic = "^1.8.1"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
